+++ services/processor/processor/__init__.py
@@ -1,0 +1,6 @@
+from .ftrack_server import main
+
+
+__all__ = (
+    "main",
+)
